libdevinfo: comparison of integers of different signs
../devinfo_retire.c:652:29: error: comparison of integers of different signs: 'int' and 'size_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
                RIO_ASSERT(dp, s - plistp < len);
                               ~~~~~~~~~~ ^ ~~~
../devinfo_retire.c:65:10: note: expanded from macro 'RIO_ASSERT'
                {if (!(x)) rio_assert(d, #x, __LINE__, __FILE__); }
                       ^
../devinfo_retire.c:656:28: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
        RIO_ASSERT(dp, s - plistp == len - 1);
                       ~~~~~~~~~~ ^  ~~~~~~~
../devinfo_retire.c:65:10: note: expanded from macro 'RIO_ASSERT'
                {if (!(x)) rio_assert(d, #x, __LINE__, __FILE__); }
                       ^

../devinfo_realpath.c:122:16: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
                if (s - left >= sizeof (next_token)) {
                    ~~~~~~~~ ^  ~~~~~~~~~~~~~~~~~~~
../devinfo_realpath.c:217:19: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
                                        if (slen + 1 >= sizeof (symlink)) {
                                            ~~~~~~~~ ^  ~~~~~~~~~~~~~~~~
2 errors generated.

../devinfo_finddev.c:145:9: error: comparison of integers of different signs: 'size_t' (aka 'unsigned int') and 'int' [-Werror,-Wsign-compare]
                if (n == handle->npaths) {
                    ~ ^  ~~~~~~~~~~~~~~
1 error generated.

../devinfo_devlink.c:497:56: error: comparison of integers of different signs: 'uint32_t' (aka 'unsigned int') and 'long' [-Werror,-Wsign-compare]
        if (DB_HDR(hdp)->page_sz == 0 || DB_HDR(hdp)->page_sz != page_sz) {
                                         ~~~~~~~~~~~~~~~~~~~~ ^  ~~~~~~~

../devinfo_devlink.c:1931:22: error: comparison of integers of different signs: 'int' and 'uint_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
        for (hval = 0; hval < CACHE(hdp)->hash_sz; hval++) {
                       ~~~~ ^ ~~~~~~~~~~~~~~~~~~~

../devinfo_devlink.c:2548:16: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
        for (i = 0; i < N_SKIP_DIRS; i++) {
                    ~ ^ ~~~~~~~~~~~
../devinfo_devlink.c:2576:17: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
                for (i = 0; i < N_SKIP_FILES; i++) {
                            ~ ^ ~~~~~~~~~~~~
../devinfo_devlink.c:2900:16: error: comparison of integers of different signs: 'int' and 'db_seg_t' (aka 'enum db_seg') [-Werror,-Wsign-compare]
        for (s = 0; s < seg; s++) {
                    ~ ^ ~~~

../devinfo_devlink.c:3830:21: error: comparison of integers of different signs: 'int' and 'debug_level_t' [-Werror,-Wsign-compare]
        if (_devlink_debug < msglevel)
            ~~~~~~~~~~~~~~ ^ ~~~~~~~~
../devinfo_devlink.c:3832:47: error: comparison of integers of different signs: 'debug_level_t' and 'int' [-Werror,-Wsign-compare]
        if ((_devlink_debug == DBG_LCK) && (msglevel != _devlink_debug))
                                            ~~~~~~~~ ^  ~~~~~~~~~~~~~~
2 errors generated.

../devinfo_prop_decode.c:974:13: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
        if (rawlen < sizeof (slots))
            ~~~~~~ ^ ~~~~~~~~~~~~~~
1 error generated.

../devinfo.c:1269:20: error: comparison of integers of different signs: 'size_t' (aka 'unsigned int') and 'int' [-Werror,-Wsign-compare]
                    (strlen(*gp) == len))
                     ~~~~~~~~~~~ ^  ~~~

../devinfo.c:3962:15: error: comparison of integers of different signs: 'int' and 'di_debug_t' [-Werror,-Wsign-compare]
        if (di_debug < msglevel)
            ~~~~~~~~ ^ ~~~~~~~~

+++ usr/src/lib/libdevinfo/devinfo_devlink.h
@@ -287,7 +287,8 @@ static struct di_devlink_handle *handle_alloc(const char *dev_dir,
     uint_t flags);
 static int cache_alloc(struct di_devlink_handle *hdp);
 static int open_db(struct di_devlink_handle *hdp, int flags);
-static int invalid_db(struct di_devlink_handle *hdp, size_t fsize, long pg_sz);
+static int invalid_db(struct di_devlink_handle *hdp, size_t fsize,
+    size_t pg_sz);
 static int read_nodes(struct di_devlink_handle *hdp, cache_node_t *pcnp,
     uint32_t nidx);
 static int read_minors(struct di_devlink_handle *hdp, cache_node_t *pcnp,
